refactor test organization (#440)
* refactor test organization

* fixed se layer

* update mmcv uper bound
