+++ mmseg/__init__.py
@@ -3,7 +3,7 @@ import mmcv
 from .version import __version__, version_info
 
 MMCV_MIN = '1.1.4'
-MMCV_MAX = '1.3.0'
+MMCV_MAX = '1.4.0'
 
 
 def digit_version(version_str):
