+++ tests/test_models/test_backbones/test_resnet.py
@@ -4,50 +4,10 @@ from mmcv.ops import DeformConv2dPack
 from mmcv.utils.parrots_wrapper import _BatchNorm
 from torch.nn.modules import AvgPool2d, GroupNorm
 
-from mmseg.models.backbones import (CGNet, FastSCNN, MobileNetV3, ResNeSt,
-                                    ResNet, ResNetV1d, ResNeXt)
-from mmseg.models.backbones.cgnet import (ContextGuidedBlock,
-                                          GlobalContextExtractor)
-from mmseg.models.backbones.resnest import Bottleneck as BottleneckS
+from mmseg.models.backbones import ResNet, ResNetV1d
 from mmseg.models.backbones.resnet import BasicBlock, Bottleneck
-from mmseg.models.backbones.resnext import Bottleneck as BottleneckX
 from mmseg.models.utils import ResLayer
-
-
-def is_block(modules):
-    """Check if is ResNet building block."""
-    if isinstance(modules, (BasicBlock, Bottleneck, BottleneckX)):
-        return True
-    return False
-
-
-def is_norm(modules):
-    """Check if is one of the norms."""
-    if isinstance(modules, (GroupNorm, _BatchNorm)):
-        return True
-    return False
-
-
-def all_zeros(modules):
-    """Check if the weight(and bias) is all zero."""
-    weight_zero = torch.allclose(modules.weight.data,
-                                 torch.zeros_like(modules.weight.data))
-    if hasattr(modules, 'bias'):
-        bias_zero = torch.allclose(modules.bias.data,
-                                   torch.zeros_like(modules.bias.data))
-    else:
-        bias_zero = True
-
-    return weight_zero and bias_zero
-
-
-def check_norm_state(modules, train_state):
-    """Check if norm layer is in correct train state."""
-    for mod in modules:
-        if isinstance(mod, _BatchNorm):
-            if mod.training != train_state:
-                return False
-    return True
+from .utils import all_zeros, check_norm_state, is_block, is_norm
 
 
 def test_resnet_basic_block():
@@ -611,329 +571,3 @@ def test_resnet_backbone():
     assert feat[1].shape == torch.Size([1, 512, 28, 28])
     assert feat[2].shape == torch.Size([1, 1024, 14, 14])
     assert feat[3].shape == torch.Size([1, 2048, 7, 7])
-
-
-def test_renext_bottleneck():
-    with pytest.raises(AssertionError):
-        # Style must be in ['pytorch', 'caffe']
-        BottleneckX(64, 64, groups=32, base_width=4, style='tensorflow')
-
-    # Test ResNeXt Bottleneck structure
-    block = BottleneckX(
-        64, 64, groups=32, base_width=4, stride=2, style='pytorch')
-    assert block.conv2.stride == (2, 2)
-    assert block.conv2.groups == 32
-    assert block.conv2.out_channels == 128
-
-    # Test ResNeXt Bottleneck with DCN
-    dcn = dict(type='DCN', deform_groups=1, fallback_on_stride=False)
-    with pytest.raises(AssertionError):
-        # conv_cfg must be None if dcn is not None
-        BottleneckX(
-            64,
-            64,
-            groups=32,
-            base_width=4,
-            dcn=dcn,
-            conv_cfg=dict(type='Conv'))
-    BottleneckX(64, 64, dcn=dcn)
-
-    # Test ResNeXt Bottleneck forward
-    block = BottleneckX(64, 16, groups=32, base_width=4)
-    x = torch.randn(1, 64, 56, 56)
-    x_out = block(x)
-    assert x_out.shape == torch.Size([1, 64, 56, 56])
-
-
-def test_resnext_backbone():
-    with pytest.raises(KeyError):
-        # ResNeXt depth should be in [50, 101, 152]
-        ResNeXt(depth=18)
-
-    # Test ResNeXt with group 32, base_width 4
-    model = ResNeXt(depth=50, groups=32, base_width=4)
-    print(model)
-    for m in model.modules():
-        if is_block(m):
-            assert m.conv2.groups == 32
-    model.init_weights()
-    model.train()
-
-    imgs = torch.randn(1, 3, 224, 224)
-    feat = model(imgs)
-    assert len(feat) == 4
-    assert feat[0].shape == torch.Size([1, 256, 56, 56])
-    assert feat[1].shape == torch.Size([1, 512, 28, 28])
-    assert feat[2].shape == torch.Size([1, 1024, 14, 14])
-    assert feat[3].shape == torch.Size([1, 2048, 7, 7])
-
-
-def test_fastscnn_backbone():
-    with pytest.raises(AssertionError):
-        # Fast-SCNN channel constraints.
-        FastSCNN(
-            3, (32, 48),
-            64, (64, 96, 128), (2, 2, 1),
-            global_out_channels=127,
-            higher_in_channels=64,
-            lower_in_channels=128)
-
-    # Test FastSCNN Standard Forward
-    model = FastSCNN()
-    model.init_weights()
-    model.train()
-    batch_size = 4
-    imgs = torch.randn(batch_size, 3, 512, 1024)
-    feat = model(imgs)
-
-    assert len(feat) == 3
-    # higher-res
-    assert feat[0].shape == torch.Size([batch_size, 64, 64, 128])
-    # lower-res
-    assert feat[1].shape == torch.Size([batch_size, 128, 16, 32])
-    # FFM output
-    assert feat[2].shape == torch.Size([batch_size, 128, 64, 128])
-
-
-def test_resnest_bottleneck():
-    with pytest.raises(AssertionError):
-        # Style must be in ['pytorch', 'caffe']
-        BottleneckS(64, 64, radix=2, reduction_factor=4, style='tensorflow')
-
-    # Test ResNeSt Bottleneck structure
-    block = BottleneckS(
-        64, 256, radix=2, reduction_factor=4, stride=2, style='pytorch')
-    assert block.avd_layer.stride == 2
-    assert block.conv2.channels == 256
-
-    # Test ResNeSt Bottleneck forward
-    block = BottleneckS(64, 16, radix=2, reduction_factor=4)
-    x = torch.randn(2, 64, 56, 56)
-    x_out = block(x)
-    assert x_out.shape == torch.Size([2, 64, 56, 56])
-
-
-def test_resnest_backbone():
-    with pytest.raises(KeyError):
-        # ResNeSt depth should be in [50, 101, 152, 200]
-        ResNeSt(depth=18)
-
-    # Test ResNeSt with radix 2, reduction_factor 4
-    model = ResNeSt(
-        depth=50, radix=2, reduction_factor=4, out_indices=(0, 1, 2, 3))
-    model.init_weights()
-    model.train()
-
-    imgs = torch.randn(2, 3, 224, 224)
-    feat = model(imgs)
-    assert len(feat) == 4
-    assert feat[0].shape == torch.Size([2, 256, 56, 56])
-    assert feat[1].shape == torch.Size([2, 512, 28, 28])
-    assert feat[2].shape == torch.Size([2, 1024, 14, 14])
-    assert feat[3].shape == torch.Size([2, 2048, 7, 7])
-
-
-def test_cgnet_GlobalContextExtractor():
-    block = GlobalContextExtractor(16, 16, with_cp=True)
-    x = torch.randn(2, 16, 64, 64, requires_grad=True)
-    x_out = block(x)
-    assert x_out.shape == torch.Size([2, 16, 64, 64])
-
-
-def test_cgnet_context_guided_block():
-    with pytest.raises(AssertionError):
-        # cgnet ContextGuidedBlock GlobalContextExtractor channel and reduction
-        # constraints.
-        ContextGuidedBlock(8, 8)
-
-    # test cgnet ContextGuidedBlock with checkpoint forward
-    block = ContextGuidedBlock(
-        16, 16, act_cfg=dict(type='PReLU'), with_cp=True)
-    assert block.with_cp
-    x = torch.randn(2, 16, 64, 64, requires_grad=True)
-    x_out = block(x)
-    assert x_out.shape == torch.Size([2, 16, 64, 64])
-
-    # test cgnet ContextGuidedBlock without checkpoint forward
-    block = ContextGuidedBlock(32, 32)
-    assert not block.with_cp
-    x = torch.randn(3, 32, 32, 32)
-    x_out = block(x)
-    assert x_out.shape == torch.Size([3, 32, 32, 32])
-
-    # test cgnet ContextGuidedBlock with down sampling
-    block = ContextGuidedBlock(32, 32, downsample=True)
-    assert block.conv1x1.conv.in_channels == 32
-    assert block.conv1x1.conv.out_channels == 32
-    assert block.conv1x1.conv.kernel_size == (3, 3)
-    assert block.conv1x1.conv.stride == (2, 2)
-    assert block.conv1x1.conv.padding == (1, 1)
-
-    assert block.f_loc.in_channels == 32
-    assert block.f_loc.out_channels == 32
-    assert block.f_loc.kernel_size == (3, 3)
-    assert block.f_loc.stride == (1, 1)
-    assert block.f_loc.padding == (1, 1)
-    assert block.f_loc.groups == 32
-    assert block.f_loc.dilation == (1, 1)
-    assert block.f_loc.bias is None
-
-    assert block.f_sur.in_channels == 32
-    assert block.f_sur.out_channels == 32
-    assert block.f_sur.kernel_size == (3, 3)
-    assert block.f_sur.stride == (1, 1)
-    assert block.f_sur.padding == (2, 2)
-    assert block.f_sur.groups == 32
-    assert block.f_sur.dilation == (2, 2)
-    assert block.f_sur.bias is None
-
-    assert block.bottleneck.in_channels == 64
-    assert block.bottleneck.out_channels == 32
-    assert block.bottleneck.kernel_size == (1, 1)
-    assert block.bottleneck.stride == (1, 1)
-    assert block.bottleneck.bias is None
-
-    x = torch.randn(1, 32, 32, 32)
-    x_out = block(x)
-    assert x_out.shape == torch.Size([1, 32, 16, 16])
-
-    # test cgnet ContextGuidedBlock without down sampling
-    block = ContextGuidedBlock(32, 32, downsample=False)
-    assert block.conv1x1.conv.in_channels == 32
-    assert block.conv1x1.conv.out_channels == 16
-    assert block.conv1x1.conv.kernel_size == (1, 1)
-    assert block.conv1x1.conv.stride == (1, 1)
-    assert block.conv1x1.conv.padding == (0, 0)
-
-    assert block.f_loc.in_channels == 16
-    assert block.f_loc.out_channels == 16
-    assert block.f_loc.kernel_size == (3, 3)
-    assert block.f_loc.stride == (1, 1)
-    assert block.f_loc.padding == (1, 1)
-    assert block.f_loc.groups == 16
-    assert block.f_loc.dilation == (1, 1)
-    assert block.f_loc.bias is None
-
-    assert block.f_sur.in_channels == 16
-    assert block.f_sur.out_channels == 16
-    assert block.f_sur.kernel_size == (3, 3)
-    assert block.f_sur.stride == (1, 1)
-    assert block.f_sur.padding == (2, 2)
-    assert block.f_sur.groups == 16
-    assert block.f_sur.dilation == (2, 2)
-    assert block.f_sur.bias is None
-
-    x = torch.randn(1, 32, 32, 32)
-    x_out = block(x)
-    assert x_out.shape == torch.Size([1, 32, 32, 32])
-
-
-def test_cgnet_backbone():
-    with pytest.raises(AssertionError):
-        # check invalid num_channels
-        CGNet(num_channels=(32, 64, 128, 256))
-
-    with pytest.raises(AssertionError):
-        # check invalid num_blocks
-        CGNet(num_blocks=(3, 21, 3))
-
-    with pytest.raises(AssertionError):
-        # check invalid dilation
-        CGNet(num_blocks=2)
-
-    with pytest.raises(AssertionError):
-        # check invalid reduction
-        CGNet(reductions=16)
-
-    with pytest.raises(AssertionError):
-        # check invalid num_channels and reduction
-        CGNet(num_channels=(32, 64, 128), reductions=(64, 129))
-
-    # Test CGNet with default settings
-    model = CGNet()
-    model.init_weights()
-    model.train()
-
-    imgs = torch.randn(2, 3, 224, 224)
-    feat = model(imgs)
-    assert len(feat) == 3
-    assert feat[0].shape == torch.Size([2, 35, 112, 112])
-    assert feat[1].shape == torch.Size([2, 131, 56, 56])
-    assert feat[2].shape == torch.Size([2, 256, 28, 28])
-
-    # Test CGNet with norm_eval True and with_cp True
-    model = CGNet(norm_eval=True, with_cp=True)
-    with pytest.raises(TypeError):
-        # check invalid pretrained
-        model.init_weights(pretrained=8)
-    model.init_weights()
-    model.train()
-
-    imgs = torch.randn(2, 3, 224, 224)
-    feat = model(imgs)
-    assert len(feat) == 3
-    assert feat[0].shape == torch.Size([2, 35, 112, 112])
-    assert feat[1].shape == torch.Size([2, 131, 56, 56])
-    assert feat[2].shape == torch.Size([2, 256, 28, 28])
-
-
-def test_mobilenet_v3():
-    with pytest.raises(AssertionError):
-        # check invalid arch
-        MobileNetV3('big')
-
-    with pytest.raises(AssertionError):
-        # check invalid reduction_factor
-        MobileNetV3(reduction_factor=0)
-
-    with pytest.raises(ValueError):
-        # check invalid out_indices
-        MobileNetV3(out_indices=(0, 1, 15))
-
-    with pytest.raises(ValueError):
-        # check invalid frozen_stages
-        MobileNetV3(frozen_stages=15)
-
-    with pytest.raises(TypeError):
-        # check invalid pretrained
-        model = MobileNetV3()
-        model.init_weights(pretrained=8)
-
-    # Test MobileNetV3 with default settings
-    model = MobileNetV3()
-    model.init_weights()
-    model.train()
-
-    imgs = torch.randn(2, 3, 224, 224)
-    feat = model(imgs)
-    assert len(feat) == 3
-    assert feat[0].shape == (2, 16, 112, 112)
-    assert feat[1].shape == (2, 16, 56, 56)
-    assert feat[2].shape == (2, 576, 28, 28)
-
-    # Test MobileNetV3 with arch = 'large'
-    model = MobileNetV3(arch='large', out_indices=(1, 3, 16))
-    model.init_weights()
-    model.train()
-
-    imgs = torch.randn(2, 3, 224, 224)
-    feat = model(imgs)
-    assert len(feat) == 3
-    assert feat[0].shape == (2, 16, 112, 112)
-    assert feat[1].shape == (2, 24, 56, 56)
-    assert feat[2].shape == (2, 960, 28, 28)
-
-    # Test MobileNetV3 with norm_eval True, with_cp True and frozen_stages=5
-    model = MobileNetV3(norm_eval=True, with_cp=True, frozen_stages=5)
-    with pytest.raises(TypeError):
-        # check invalid pretrained
-        model.init_weights(pretrained=8)
-    model.init_weights()
-    model.train()
-
-    imgs = torch.randn(2, 3, 224, 224)
-    feat = model(imgs)
-    assert len(feat) == 3
-    assert feat[0].shape == (2, 16, 112, 112)
-    assert feat[1].shape == (2, 16, 56, 56)
-    assert feat[2].shape == (2, 576, 28, 28)
